Updated image preview support

+++ preview.php
@@ -29,22 +29,17 @@ include "./utils.php";
         $raw_instance_configuration = file_get_contents($instance_configuration_file);
         $instance_config = json_decode($raw_instance_configuration, true);
 
+        $image_file = $config["image_stream"];
 
-        if (isset($instance_config["developer"]["working_directory"])) { // Check to see if the configuration layout reflects that the instance Predator Fabric.
-            $image_file = $instance_config["developer"]["working_directory"] . "/" . $instance_config["image"]["camera"]["file_name"];
-        } else if (isset($instance_config["general"]["working_directory"])) { // Check to see if the configuration layout reflects that the instance is vanilla Predator.
-            $image_file = $instance_config["general"]["working_directory"] . "/" . $instance_config["realtime"]["image"]["camera"]["file_name"] . ".jpg";
+        if (file_exists($image_file) == true) { // Check to make sure the image file exists before displaying it.
+            $image_data = fread(fopen($image_file, "r"), filesize($image_file));
+
+            echo "<div style=\"width:100%;text-align:center;\">";
+            echo "    <img style=\"max-width:100%;max-height:680px;\" src='" . 'data:image;base64,' . base64_encode($image_data) . "' />";
+            echo "</div>";
         } else {
-            echo "<p class=\"error\">The image file could not be located given the instance configuration file.</p>";
+            echo "<p><i>No image available.</i></p>";
         }
-
-        $image_data = fread(fopen($image_file, "r"), filesize($image_file));
-
-        echo "<div style=\"width:100%;text-align:center;\">";
-        echo "    <img style=\"max-width:100%;max-height:680px;\" src='" . 'data:image;base64,' . base64_encode($image_data) . "' />";
-        echo "</div>";
-
-
         ?>
     </body>
 </html>
